requirements-test.txt
pylint
django
djangorestframework
factory-boy
